refactor(cas): mv graph to model
